third_party/blink/web_tests/fast/xpath/node-name-case-sensitivity-expected.txt

strongstrongFOOfoo

HTML //*[@id="sometext"]//strong
PASS res.snapshotLength is 2
HTML //*[@id="sometext"]//Strong
PASS res.snapshotLength is 2
HTML //*[@id="Sometext"]//strong
PASS res.snapshotLength is 0
HTML //*[@id="sometext"]//foo
PASS res.snapshotLength is 2
HTML //*[@id="sometext"]//FOO
PASS res.snapshotLength is 2
XML //*[@id="sometext"]//x:strong
PASS res.snapshotLength is 1
XML //*[@id="sometext"]//x:Strong
PASS res.snapshotLength is 0
XML //*[@id="Sometext"]//x:strong
PASS res.snapshotLength is 0
XML //*[@id="sometext"]//foo
PASS res.snapshotLength is 1
XML //*[@id="sometext"]//FOO
PASS res.snapshotLength is 1
PASS successfullyParsed is true

TEST COMPLETE
